requirements.txt
wheel
pandas 
tensorflow-macos
tensorflow-metal
matplotlib
sckit-learn
seaborn
diskcache
statsmodels